more thoughts on rates

+++ src/time/daycounters.rs
@@ -7,5 +7,5 @@ use chrono::NaiveDate;
 
 pub trait DayCounters {
     fn day_count(&self, d1: NaiveDate, d2: NaiveDate) -> i64;
-    fn year_fraction(&self, d1: NaiveDate, d2: NaiveDate) -> f32;
+    fn year_fraction(&self, d1: NaiveDate, d2: NaiveDate) -> f64;
 }
